Reduce size of Action enum to u8
I know that this uses c_int to keep it exact to GLFW... But this is
just a waste of bits.

+++ src/action.zig
@@ -1,7 +1,7 @@
 const c = @import("c.zig").c;
 
 /// Key and button actions
-pub const Action = enum(c_int) {
+pub const Action = enum(u8) {
     /// The key or mouse button was released.
     release = c.GLFW_RELEASE,
